requirements.txt
libarchive-c
rarfile
requests
tqdm
libtorrent
